+++ backend/core/src/auth/auth.module.ts
@@ -23,6 +23,8 @@ import { SmsMfaService } from "./services/sms-mfa.service"
 import { TwilioModule } from "nestjs-twilio"
 import { Listing } from "../listings/entities/listing.entity"
 import { ListingsModule } from "../listings/listings.module"
+import { UserCsvExporterService } from "./services/user-csv-exporter.service"
+import { CsvBuilder } from "../applications/services/csv-builder.service"
 
 @Module({
   imports: [
@@ -60,6 +62,8 @@ import { ListingsModule } from "../listings/listings.module"
     UserService,
     PasswordService,
     SmsMfaService,
+    CsvBuilder,
+    UserCsvExporterService,
   ],
   exports: [AuthzService, AuthService, UserService],
   controllers: [AuthController, UserController, UserProfileController],