+++ backend/core/test/activity-logs/activity-log.e2e-spec.ts
@@ -29,7 +29,7 @@ import { getUserAccessToken } from "../utils/get-user-access-token"
 declare const expect: jest.Expect
 jest.setTimeout(30000)
 
-describe("Activiy", () => {
+describe("Activity", () => {
   let app: INestApplication
   let adminId: string
   let adminAccessToken: string